JIT Compilation (#54)
Changes the implementation of scheme-rs from an interpreter to a JIT compiler.

Most features are unharmed, but a few need to be re-implemented. dynamic-wind is one example. This rearchitecture will also make some of these, along with features that were not implemented before like exception handlers, much easier to implement. 

Performance is impacted quite a bit, but this rearchitecture was necessary to ensure competitive performance is achievable at all. We'll need to add analysis and passes to our compiler, we had to have CPS in order to perform the analysis.

Bugs are also probably expected.