Implement `dynamic-wind` (#89)
Implements `dynamic-wind` control structure, by passing a dynamic extent (set of in- and out-winders) through the execution trampoline. 

This implementation is largely functional _except_ for raising in a dynamic wind, which fails to call the out winders. That will be added as a separate issue and fixed at a later time.